made new Helper class for independent functions & put dayOfTheYear func there
